+++ rumqttc/src/tls.rs
@@ -49,11 +49,14 @@ pub enum Error {
     #[error("TLS error: {0}")]
     TLS(#[from] rustls::Error),
     #[cfg(feature = "use-rustls")]
-    /// No valid certificate in chain
+    /// No valid CA cert found
     #[error("No valid CA certificate provided")]
-    NoValidCACert,
-    #[error("No valid certificate in chain")]
     NoValidCertInChain,
+    #[cfg(feature = "use-rustls")]
+    /// No valid client cert found
+    #[error("No valid certificate for client authentication in chain")]
+    NoValidClientCertInChain,
+    #[cfg(feature = "use-rustls")]
     /// No valid key found
     #[error("No valid key in chain")]
     NoValidKeyInChain,
@@ -61,13 +64,6 @@ pub enum Error {
     #[error("Native TLS error {0}")]
     NativeTls(#[from] NativeTlsError),
 }
-
-// // The cert handling functions return unit right now, this is a shortcut
-// impl From<()> for Error {
-//     fn from(_: ()) -> Self {
-//         Error::NoValidCertInChain
-//     }
-// }
 
 #[cfg(feature = "use-rustls")]
 pub async fn rustls_connector(tls_config: &TlsConfiguration) -> Result<RustlsConnector, Error> {
@@ -98,7 +94,7 @@ pub async fn rustls_connector(tls_config: &TlsConfiguration) -> Result<RustlsCon
             root_cert_store.add_trust_anchors(trust_anchors);
 
             if root_cert_store.is_empty() {
-                return Err(Error::NoValidCACert);
+                return Err(Error::NoValidCertInChain);
             }
 
             let config = ClientConfig::builder()
@@ -107,14 +103,11 @@ pub async fn rustls_connector(tls_config: &TlsConfiguration) -> Result<RustlsCon
 
             // Add der encoded client cert and key
             let mut config = if let Some(client) = client_auth.as_ref() {
-
                 let certs =
                     rustls_pemfile::certs(&mut BufReader::new(Cursor::new(client.0.clone())))?;
                 if certs.is_empty() {
-                    return Err(Error::NoValidCertInChain);
+                    return Err(Error::NoValidClientCertInChain);
                 }
-                // Load appropriate Key as per the user request. The underlying signature algorithm
-                // of key generation determines the Signature Algorithm during the TLS Handskahe.
 
                 // Create buffer for key file
                 let mut key_buffer = BufReader::new(Cursor::new(client.1.clone()));