+++ rumqttc/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Expose `EventLoop::clean` to allow triggering shutdown and subsequent storage of pending requests
 - Support for all variants of TLS key formats currently supported by Rustls: `PKCS#1`, `PKCS#8`, `RFC5915`. In practice we should now support all RSA keys and ECC keys in `DER` and `SEC1` encoding. Previously only `PKCS#1` and `PKCS#8` where supported.
-- TLS Error variants: `NoValidCACert`, `NoValidKeyInChain`.
+- TLS Error variants: `NoValidClientCertInChain`, `NoValidKeyInChain`.
 
 ### Changed
 - Synchronous client methods take `&self` instead of `&mut self` (#646)